+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Support disk usage reporting (using `df`) on Alpine Linux.
+- When a disk mountpoint has no inodes usage percentage, skip the mountpoint, and report the inodes information successfully for the inodes that do have an inodes usage percentage.
+
 ## 0.5.2
 
 - Normalize CPU usage percentages for cgroups v2 systems